no need to call toInt on an Int

+++ library/src/main/scala/request/headers.scala
@@ -152,7 +152,7 @@ object HostPort {
   def unapply[T](req: HttpRequest[T]): Option[(String, Int)] =
     req match {
       case Host(hostname) => hostname.split(':') match {
-        case Array(host, Of.Int(port)) => Some(host, port.toInt)
+        case Array(host, Of.Int(port)) => Some(host, port)
         case _ => Some(hostname, if(req.isSecure) 443 else 80)
       }
       case _ => None
